requirements.txt
requests
python-dotenv
openai
newspaper3k
supabase
schedule
tweepy
urllib3
certifi
charset-normalizer
idna
feedparser
lxml_html_clean
cloudinary
huggingface_hub
streamlit
